# Check target array length before allocating.