--- a/core/jvm-native/src/main/scala/cats/effect/ArrayStack.scala
+++ b/core/jvm-native/src/main/scala/cats/effect/ArrayStack.scala
@@ -17,6 +17,7 @@
 package cats.effect
 
 import Platform.static
+import PlatformStatics.VM_MaxArraySize
 
 private final class ArrayStack[A <: AnyRef](
     private[this] var buffer: Array[AnyRef],
@@ -72,7 +73,17 @@ private final class ArrayStack[A <: AnyRef](
   private[this] def checkAndGrow(): Unit =
     if (index >= buffer.length) {
       val len = buffer.length
-      val buffer2 = new Array[AnyRef](len * 2)
+      val targetLen = len * 2
+
+      val resizeLen =
+        if (targetLen < 0)
+          throw new Exception(s"Overflow while resizing array. Request length: $targetLen")
+        else if (len > VM_MaxArraySize / 2)
+          VM_MaxArraySize
+        else
+          targetLen
+
+      val buffer2 = new Array[AnyRef](resizeLen)
       System.arraycopy(buffer, 0, buffer2, 0, len)
       buffer = buffer2
     }