--- a/core/jvm-native/src/main/scala/cats/effect/ArrayStack.scala
+++ b/core/jvm-native/src/main/scala/cats/effect/ArrayStack.scala
@@ -17,7 +17,6 @@
 package cats.effect
 
 import PlatformStatics.VM_MaxArraySize
-
 import Platform.static
 
 private[effect] final class ArrayStack[A <: AnyRef](