--- a/core/jvm-native/src/main/scala/cats/effect/CallbackStack.scala
+++ b/core/jvm-native/src/main/scala/cats/effect/CallbackStack.scala
@@ -20,8 +20,7 @@ import scala.annotation.tailrec
 
 import java.util.concurrent.atomic.{AtomicBoolean, AtomicReference}
 
-import CallbackStack.Handle
-import CallbackStack.Node
+import CallbackStack.{Handle, Node}
 import Platform.static
 
 private final class CallbackStack[A](private[this] var callback: A => Unit)