--- a/core/jvm-native/src/main/scala/cats/effect/PlatformStatics.scala
+++ b/core/jvm-native/src/main/scala/cats/effect/PlatformStatics.scala
@@ -1,0 +1,8 @@
+package cats.effect
+
+// copy from scala.runtime.PStatics
+private[effect] object PlatformStatics {
+  // `Int.MaxValue - 8` traditional soft limit to maximize compatibility with diverse JVMs
+  // See https://stackoverflow.com/a/8381338 for example
+  final val VM_MaxArraySize = 2147483639
+}