The current implementation of the array inside `ArrayStack` does not verify the length of the new array, and this may lead to overflow problem.

https://github.com/typelevel/cats-effect/blob/ecf93db298c7fdb33df3be96dbe24e8f9bf90b9c/core/jvm-native/src/main/scala/cats/effect/ArrayStack.scala#L72-L78

Here we add a check on
- Int overflow
- Array length exceed VM limit

see also: https://github.com/typelevel/cats-effect/pull/4135