@armanbilge 2024-10-29
Thanks for the PR! See also:
- https://github.com/typelevel/cats-effect/pull/3949#issuecomment-1913198330