@djspiewak 2024-10-29
Looks like we need to run `prePR`. But apart from that, I'd like to sanity check this with a benchmark run. This *specific* function is very, very easy to knock out of its JIT fastpath, and adding this type of branching very much runs that risk. Apart from that, I think this is a very good change. If we are knocking the JIT around, we might have to either experiment with different ways of encoding the branching to avoid tripping over the optimizer, or we may need to do the check in some other area.