@djspiewak 2024-11-21
`sbt clean prePR` might do the trick on the build. Scalafix and scalafmt both have some weird caching issues.

I'll run some benchmarks on this via `IO` when a get a chance so we can see how it impacts inlining. The `ArrayStack` direct results are promising.