@djspiewak 2024-12-08
Just when I think the JVM can't surprise me anymore...

## Before

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score     Error    Units
[info] DeepBindBenchmark.async                                       N/A    10000  thrpt   10  2648.872 ±  71.889    ops/s
[info] DeepBindBenchmark.delay                                       N/A    10000  thrpt   10  7472.865 ± 263.016    ops/s
[info] DeepBindBenchmark.pure                                        N/A    10000  thrpt   10  8595.472 ± 256.137    ops/s
[info] MapCallsBenchmark.batch120                                    N/A      N/A  thrpt   10   187.446 ±   1.039    ops/s
[info] MapCallsBenchmark.batch30                                     N/A      N/A  thrpt   10    47.649 ±   1.047    ops/s
[info] MapCallsBenchmark.one                                         N/A      N/A  thrpt   10     1.597 ±   0.021    ops/s
[info] MapStreamBenchmark.batch120                                   N/A      N/A  thrpt   10  5017.096 ±  24.569    ops/s
[info] MapStreamBenchmark.batch30                                    N/A      N/A  thrpt   10  2411.243 ±  49.510    ops/s
[info] MapStreamBenchmark.one                                        N/A      N/A  thrpt   10  2883.050 ±  21.626    ops/s
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   747.384 ±  37.987    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    73.515 ±   1.896    ops/s
[info] ShallowBindBenchmark.async                                    N/A    10000  thrpt   10  1939.453 ±  48.402    ops/s
[info] ShallowBindBenchmark.delay                                    N/A    10000  thrpt   10  7428.462 ±  25.963    ops/s
[info] ShallowBindBenchmark.pure                                     N/A    10000  thrpt   10  8098.391 ±  53.194    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    12.481 ±   0.114  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    28.263 ±   5.321  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2503.856 ± 157.461  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  1957.221 ±  65.798  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    28.080 ±   3.298  ops/min
```

## After

```
[info] Benchmark                                             (cpuTokens)   (size)   Mode  Cnt     Score     Error    Units
[info] DeepBindBenchmark.async                                       N/A    10000  thrpt   10  2668.144 ± 114.274    ops/s
[info] DeepBindBenchmark.delay                                       N/A    10000  thrpt   10  7552.944 ± 179.589    ops/s
[info] DeepBindBenchmark.pure                                        N/A    10000  thrpt   10  8691.767 ±  39.296    ops/s
[info] MapCallsBenchmark.batch120                                    N/A      N/A  thrpt   10   191.787 ±   2.813    ops/s
[info] MapCallsBenchmark.batch30                                     N/A      N/A  thrpt   10    48.432 ±   0.281    ops/s
[info] MapCallsBenchmark.one                                         N/A      N/A  thrpt   10     1.631 ±   0.036    ops/s
[info] MapStreamBenchmark.batch120                                   N/A      N/A  thrpt   10  5206.007 ± 191.124    ops/s
[info] MapStreamBenchmark.batch30                                    N/A      N/A  thrpt   10  2414.824 ±  66.890    ops/s
[info] MapStreamBenchmark.one                                        N/A      N/A  thrpt   10  2986.513 ±  16.887    ops/s
[info] ParallelBenchmark.parTraverse                               10000     1000  thrpt   10   780.594 ±   4.945    ops/s
[info] ParallelBenchmark.traverse                                  10000     1000  thrpt   10    73.811 ±   0.322    ops/s
[info] ShallowBindBenchmark.async                                    N/A    10000  thrpt   10  1980.663 ±  78.412    ops/s
[info] ShallowBindBenchmark.delay                                    N/A    10000  thrpt   10  7543.541 ± 285.638    ops/s
[info] ShallowBindBenchmark.pure                                     N/A    10000  thrpt   10  8130.712 ±  21.583    ops/s
[info] WorkStealingBenchmark.alloc                                   N/A  1000000  thrpt   10    12.626 ±   0.110  ops/min
[info] WorkStealingBenchmark.manyThreadsSchedulingBenchmark          N/A  1000000  thrpt   10    35.189 ±   4.395  ops/min
[info] WorkStealingBenchmark.runnableScheduling                      N/A  1000000  thrpt   10  2715.206 ± 115.029  ops/min
[info] WorkStealingBenchmark.runnableSchedulingScalaGlobal           N/A  1000000  thrpt   10  2034.349 ±  40.435  ops/min
[info] WorkStealingBenchmark.scheduling                              N/A  1000000  thrpt   10    28.354 ±   2.660  ops/min
```